requirements.txt
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
pydantic
pyopenssl>=23.0.0
uvicorn
